requirements.txt
aiofiles==24.1.0
aiohttp==3.10.3
asyncio==3.4.3
httpx==0.25.0
motor==3.6.0
ntgcalls==1.3.3
numpy==2.2.5
pillow==9.5.0
py-tgcalls==2.1.1
python-dotenv==1.0.1
tgcrypto==1.2.5
google-generativeai
https://github.com/adityahalder/pyrogram/archive/master.zip
